Refactored cellpose into external example API and added flexibility for user to add their own channel + API directly into chatbot
